edit files

+++ backend/controllers/stuff.js
@@ -1,12 +1,15 @@
+const thing = require('../models/thing');
 const Thing = require('../models/thing');
 
 exports.createThing = (req, res, next) => {
+  req.body.thing = JSON.parse(req.body.thing);
+  const url = req.protocol + '://' + req.get('host');
   const thing = new Thing({
-    title: req.body.title,
-    description: req.body.description,
-    imageUrl: req.body.imageUrl,
-    price: req.body.price,
-    userId: req.body.userId,
+    title: req.body.thing.title,
+    description: req.body.thing.description,
+    imageUrl: url + '/images/' + req.file.filename,
+    price: req.body.thing.price,
+    userId: req.body.thing.userId,
   });
   thing
     .save()
@@ -59,17 +62,29 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: 'Deleted!',
+  Thing.findOne({ _id: req.params.id }).then((thing) => {
+    if (!thing) {
+      return res.status(404).json({
+        error: new Error('No such Thing!'),
       });
-    })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
+    }
+    if (thing.userId !== req.auth.userId) {
+      return res.status(401).json({
+        error: new Error('Unauthorized request!'),
       });
-    });
+    }
+    Thing.deleteOne({ _id: req.params.id })
+      .then(() => {
+        res.status(200).json({
+          message: 'Deleted!',
+        });
+      })
+      .catch((error) => {
+        res.status(400).json({
+          error: error,
+        });
+      });
+  });
 };
 
 exports.getAllStuff = (req, res, next) => {
